api/requirements.txt
Flask==2.2.5
Werkzeug==2.2.2
MarkupSafe==2.1.1
scikit-learn==1.5.2
tensorflow==2.18.0
gunicorn==21.2.0
flask-cors==3.0.10
numpy==2.0.1
pillow==11.0.0
